LayoutTests/platform/chromium-cg-mac/plugins/mouse-events-expected.txt

CONSOLE MESSAGE: line 18: PLUGIN: getFocusEvent
CONSOLE MESSAGE: line 18: PLUGIN: mouseDown at (12, 12)
CONSOLE MESSAGE: line 19: PLUGIN: mouseUp at (12, 12)
CONSOLE MESSAGE: line 21: PLUGIN: mouseDown at (22, 22)
CONSOLE MESSAGE: line 23: PLUGIN: mouseUp at (32, 22)

Test for bug 11517: Flash clicks/interactivity not working properly.
